+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/application/VerifyImmunizationsLoadedCommandLineRunnerTest.java
@@ -8,43 +8,30 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.dao.DataAccessException;
 import uk.nhs.adaptors.connector.dao.ImmunizationSnomedCTDao;
 
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-public class ApplicationCommandLineRunnerTest {
+public class VerifyImmunizationsLoadedCommandLineRunnerTest {
 
     @Mock
     private ImmunizationSnomedCTDao immunizationSnomedCTDao;
 
     @InjectMocks
-    private ApplicationCommandLineRunner applicationCommandLineRunner;
+    private VerifyImmunizationsLoadedCommandLineRunner verifyImmunizationsLoadedCommandLineRunner;
 
     @Test
     public void When_RunApplicationAndImmunizationsAreLoaded_Expect_ExceptionIsNotThrown() {
         when(immunizationSnomedCTDao.areImmunizationCodesLoaded()).thenReturn(true);
 
-        assertDoesNotThrow(() -> applicationCommandLineRunner.run());
+        assertDoesNotThrow(() -> verifyImmunizationsLoadedCommandLineRunner.run());
     }
 
     @Test
-    public void When_RunApplicationAndImmunizationsNotLoaded_Expect_DataAccessExceptionThrown() {
+    public void When_RunApplicationAndImmunizationsNotLoaded_Expect_RuntimeExceptionThrown() {
         when(immunizationSnomedCTDao.areImmunizationCodesLoaded()).thenReturn(false);
 
-        assertThrows(DataAccessException.class, () -> applicationCommandLineRunner.run());
+        assertThrows(RuntimeException.class, () -> verifyImmunizationsLoadedCommandLineRunner.run());
     }
-
-
-
-
-
-
-
-
-
-
-
-
 }